Remove the KeyboardInterrupt from the parents of InterruptTest.
In the source code of python in Lib/unittest/case.py, before handling
the SkipTest exception, the KeyboardInterrupt exception is handled and
if we find the latter exception, it will be raised and the test is not
skipped.  I did not find any reasons for this hierarchy and have
dropped it.

With this commit, the test_sigint passes and we are now able to skip
tests with ctrl+c.