PyCurl: Handle temporary redirect when uploading a file.
It is convenient to issue a single upload and have the files uploaded even if
the server responds with e temporary redirect.  However, we should note that
this is very easy to misuse.  The main sender might not want to receive the
actual payload.  If that is the case, we should not send the payload in the
first request.