JobCoordinator: Retry the task submission for a limited number of failures.
There is a lot of contension on the disco server so the task submission may
timeout.  We should not use large timeouts for submitting the tasks because
meanwhile the job coordinator process is blocked.
A benefit of smaller timeout is that job_coordinator can still process the
messages from other processes.