job_coordinator: Make sure a stage has finished before sending an event.
In the old model, the do_stage_done function is called only when a stage
finishes and there is no need for double checks.  However, in the concurrent
model, this function might be called when some tasks are finished.  Disco has to
make sure this stage can be finished before actually finishing it.