job_coordinator: add the tasks which are scheduled for re_run to the pending list.
One of the benefits of this approach is that when a new input is available for
this task, we know that it is not finished yet and needs the task spec to be
updated.  Otherwise, the task will be re-launched with all of the inputs but
still waiting for more inputs.