jobpack: Add the concurrent flag to the pipeline stages.
The stage() type has been modified to include a third field showing whether the
stage should run concurrently or sequentially.  By default, all of the stages
should run sequentially but it will be changed into a configurable option so
that it can be turned on for specific stages in the pipeline model.